requirements.txt
Django==1.8.9
Django-tinymce-filebrowser==0.2.1
MySQL-python==1.2.5
Pillow==2.4.0
Unidecode==0.04.17
argparse==1.2.1
beautifulsoup4==4.3.2
dj-stripe==0.3.5
django-adminfiles==1.0.1
django-ajax-selects==1.4.2
django-appconf==1.0.1
django-braces==1.4.0
django-ckeditor==4.4.7
django-ckeditor-updated==4.4.4
django-endless-pagination==2.0
django-grappelli==2.7.3
django-haystack==2.4.1
django-jsonfield==0.9.13
django-model-utils==2.0.3
django-select2-forms==1.1.25
django-shortcodes==0.3.0
django-social-share==0.2.1
django-staticfiles==1.2.1
django-tinymce==2.2.0
django-wysiwyg-redactor==0.4.9
pysolr==3.3.0
pytz==2014.10
requests==2.3.0
six==1.7.3
sorl-thumbnail==12.2
stripe==1.18.0
unicode-slugify==0.1.3
wsgiref==0.1.2
xlwt==0.7.2
djangorestframework==3.3.3